refactor route navigation
