+++ src/components/route-browser/index.js
@@ -1,7 +1,6 @@
 export * from './breadcrumbs'
 export * from './context'
-export * from './navigation-buttons'
-export * from './navigation-slider'
+export * from './navigation'
 export * from './prediction'
 export * from './scene-metadata'
 export * from './scene-prefetch'
