--- src/components/route-browser/navigation-buttons.css
@@ -1,11 +1,0 @@
-.route-navigation-button-container {
-  display: flex;
-  justify-content: center;
-  gap: 1rem;
-}
-
-.route-navigation-button {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
